Add -g options to run.sh

+++ run.sh
@@ -2,18 +2,19 @@
 
 #
 # -f オプションはダウンロードサイトに新ファイルがなくてもコンバートする
+# -g Dockerfileをgit cloneしない (デバッグ時に利用する)
 # -P オプションは並列実行プロセス数
 #
-while getopts fP: OPT
+while getopts fgP: OPT
 do
   case $OPT in
      f) OPTION_f="-f" ;;
+     g) OPTION_g="-g" ;;
      P) OPTION_P="-P$OPTARG" ;;
   esac
 done
 
 shift $(($OPTIND - 1))
-
 
 # RDF化対象のデータセット名 
 DATASET=$1
@@ -37,10 +38,10 @@ TARGET_DATASETS['pubtator']=true
 #  データセット一覧に含まれているかチェック
 #
 if [ ${#DATASET} == 0 ]; then
-  echo "Usage: run.sh [DATASET]"
+  echo "Usage: run.sh [-f] [-P number of threads] DATASET"
   exit 0
 elif ! test "${TARGET_DATASETS[$DATASET]+isset}"; then
-  echo "Usage: run.sh [DATASET]"
+  echo "Usage: run.sh [-f] [-P number of threads] DATASET"
   exit 0
 fi  
 
@@ -52,17 +53,21 @@ WORKDIR_DOWNLOAD="${WORKDIR_ROOT}/rdf-${DATASET}_download"
 #
 #  dockerファイルをgithubからclone/pullする
 #
-mkdir -p $WORKDIR_ROOT
-cd $WORKDIR_ROOT
-rm -rf $WORKDIR
-git clone https://github.com/biosciencedbc/rdf-${DATASET}
-cd "$WORKDIR"
-#
-# gitのサブモジュールを最新に更新する方法
-#
-git submodule update --recursive --init
-git submodule foreach git pull origin master
+if [ "${OPTION_g}" = "-g" ] &&  [ -d "${WORKDIR}" ]; then
+  echo "Skip git clone https://github.com/biosciencedbc/rdf-${DATASET}"
+else
+  mkdir -p $WORKDIR_ROOT
+  cd $WORKDIR_ROOT
+  rm -rf $WORKDIR
+  git clone https://github.com/biosciencedbc/rdf-${DATASET}
 
+  cd "$WORKDIR"
+  #
+  # gitのサブモジュールを最新に更新する
+  #
+  git submodule update --recursive --init
+  git submodule foreach git pull origin master
+fi
 
 # docker imageのビルド
 docker rmi rdf-${DATASET} 
@@ -79,4 +84,3 @@ mkdir -p $OUTDIR
 #  　docker stop/killでサブプロセスも含めて綺麗に停止できそう。stopもSIGTERMでなくSIGKILLで止めているようなのでkillの方が速く止まる
 #
 nohup docker run --rm -v ${WORKDIR_DOWNLOAD}:/work -v ${OUTDIR}:/data --name "rdf-${DATASET}-${YYYYMMDD}" rdf-${DATASET} ${OPTION_f} ${OPTION_P} 1> ${OUTDIR}/stdout.log  2> ${OUTDIR}/stderr.log &
-
